修改后端接收前端数据的形式

+++ blog-server/blog-server-controller/src/main/java/com/aklry/controller/AdminController.java
@@ -57,7 +57,7 @@ public class AdminController {
     @PostMapping("/delete")
     public Result deleteById(@RequestBody Integer id) {
         result = new Result();
-        adminService.deleteById(id);
+        adminService.deleteById(id.intValue());
 
         result.flag = true;
         result.message = "删除成功";
